feat(iclisten-modules): Added serialization and deserialization for iclisten modules.
refactor(CommunicationResponse): Updated tagProcessor initialization from Map.of to Map.ofEntries

refactor(PamModule): Converted PamModule to an interface instead of an abstract class

+++ src/main/java/com/acousea/backend/core/communicationSystem/application/command/DTO/NodeDeviceDTO.java
@@ -507,9 +507,9 @@ public class NodeDeviceDTO {
         public static class StatusDto {
             private Integer unitStatus;
             private Integer batteryStatus;
-            private Double batteryPercentage;
-            private Double temperature;
-            private Double humidity;
+            private Float batteryPercentage;
+            private Float temperature;
+            private Float humidity;
             private LocalDateTime timestamp;
 
             public static StatusDto fromStatus(ICListenStatus status) {
